Fixat lite
Experimente

+++ Bakdelar_API/ViewModels/ProductView.cs
@@ -15,22 +15,22 @@ namespace Bakdelar_API.ViewModels
         public string ProductDescription { get; set; }
 
         public decimal ProductPrice { get; set; }
-        
+
         public int? AvailableQuantity { get; set; }
-        
+
         public double? ProductWeight { get; set; }
         public DateTime DateEntered { get; set; }
-        public double? SpecialPrice { get; set; }
+        public decimal? SpecialPrice { get; set; }
         public bool IsSelected { get; set; }
         public int NumberOfSold { get; set; }
 
         public int CategoryId { get; set; }
-        
+
         public CategoryView Category { get; set; }
 
         public List<ProductImageView> ProductImageView { get; set; }
 
-        
+
